fixed call of overloaded entry_int is ambiguous

+++ src/json_formatter.cpp
@@ -108,7 +108,7 @@ void json_formatter::write_common(const element_info &elem) {
   writer->object_key("version");
   writer->entry_int(elem.version);
   writer->object_key("changeset");
-  writer->entry_int(elem.changeset);
+  writer->entry_int((int) elem.changeset);
   writer->object_key("timestamp");
   writer->entry_string(elem.timestamp);
   if (elem.display_name && elem.uid) {
